requirements.txt
speechrecognition==3.14.3
pyttsx3==2.71
pywhatkit==5.4
google-generativeai==0.8.5
pyaudio==0.2.14